requirements.txt
flask
flask_sqlalchemy
sqlalchemy
alembic
pillow
openai
python-dotenv
